Scrolling applications keep all the content in a single index html file. I've moved all the views into the main html file. Ensured the css is still being applied as it's ment to.
Right now all the conent is appearing in a row which is resulting in only the projects and education, experince section centent showing with the contact me to the side. I've decided to use sections instead of divs becasue the sections should appear in the file outline for easy editing and viewing.