Given that all the content is now in one index file, i can use the id names of each section to navigate to them. This also allows users to use the browser back an forth errors which keeps the history. The more you know :)
The website is now set up to what is currently being dispalyed on the site but in one page.

I've seperated out the custom section css just for later purposes. All I need now is to creat a nice transition between each section or maybe that is something I should wait for futher styling later.