Created a working version of the project section.
Started a new version of it inorder to optimize how it works.

Currenlty the styling of the next buttons are complete, as well as the photo image and project info.

I now need to attach the javascript, as well as create an option that cycles through the projects so a user can keep track of where they are.