Changed styling around social links, so they are surronded by circles.
The about me links where updated so that there are the common 7C9395 color and are surronded by a circle of the same color.

I then matched the coloring on the contact section with the other links and added text for the user to know what they are for.

Updated projects pragination so it was more accessible and changed some styling.

I still need to finish education and experince.